+++ doc/src/changelog.md
@@ -5,6 +5,7 @@ Version History
 -   Changed normal oriented curve definition to fix waving artefacts.
 -   Fixed bounding issue for normal oriented motion blurred curves.
 -   Fixed performance issue with motion blurred point geometry.
+-   Fixed generation of documentation with new pandoc versions.
 
 ### New Features in Embree 3.4.0
 -   Added point primitives (spheres, ray-oriented discs, normal-oriented discs).